requirements.txt
pandas
scikit-learn
joblib
python-whois
flask